flow-python/examples/warehouse/detection_memd/requirements.txt
onnxruntime-gpu==1.2.0
opencv-python 
numpy 
loguru
Cython